rag_system/requirements.txt
pypdf
openai
tqdm
python-dotenv
requests
transformers>=4.41.0
accelerate>=0.30.0
safetensors>=0.4.2
huggingface-hub>=0.23.0

